requirements.txt
wheel
flask
flask-restful
flask-sqlalchemy
flask-jwt-extended